==> code/analysis/03_analyzing_relative_body_mass.Rmd <==
---
title: "Analyzing body mass of recaptures"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Goal**

Analyze how experimental *Heliconia* removal influences:

- The body mass of hummingbirds captured during both experimental periods (pre and post; 'recaptures')

Predict that, if birds stayed in the treatment focal areas despite our major calorie depletion, they might lose body mass.

**Problem**

Problem with mass as a response variable: how do I account for differences in structural size within and between species? There are lots of different hummingbird species captured in this study system! Larger birds will have larger changes in body mass.

Potential options considered and ultimately not chosen:

- Random effect of species (for analysis of all species)

- Fixed effect of species (for analysis of *Heliconia* specialists, since there are only two species; not enough levels for a random effect)

- Covariate for structural size. Wing length? Wing length squared? Wing length cubed? Adding covariate for structural size is a bit tricky, because I don't know what the allometric exponent should be.

**Approach**

Analyze hummingbird body mass *relative to* mass predicted for a given wing length (within a species).

1. Using data from all years of hummingbird captures (2018-2018), create allometric equations for each hummingbird species: log(mass) ~ log(wing length)

2. For each recaptured hummingbird from the experiment (2016-2018), use the allometric equation to calculate predicted body mass for a given wing length. For each bird, I therefore need data about wing length and body mass.

3. Calculate relative mass as measured body mass divided by predicted body mass. Birds with values <1 are lighter than expected for their structural size; similarly, birds with values >1 are heavier than expected for their structural.

4. Analyze how *Heliconia* removal influences change in relative body mass.

```{r setup, include=FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE)

library(dplyr)
library(tidyr)
library(lubridate)
library(ggplot2)
library(purrr)
library(glmmTMB)
library(emmeans)
```

## Step 1: Import data and functions

**Import data**
```{r}
#Morphological measurements (mass and wing length) from capture data
data_mass_wing_regressions <- read.csv("../../data/export/for_analysis/Capture_morphological_measurements.csv") %>%
  mutate(year_patch = paste(year, patch, sep = "_")) %>%
  mutate(exp_phase = factor(exp_phase, levels = c("capture_1", "capture_2"), labels = c("pre", "post"))) %>%
  mutate(control_treatment = factor(control_treatment, levels = c("control", "treatment"))) %>%
  mutate(log_mass = log(mass)) %>%
  mutate(log_wing_length = log(wing_length))

#Resource data. Need to add quantitative treatment measure here (*Heliconia* calories removed/ha)
heto_removed <- read.csv("../../data/export/for_analysis/Site_and_replicate_characteristics.csv") %>%
  select(year, patch, calories_removed_per_ha) %>%
  mutate(patch = as.character(patch))
```

Create recapture dataset for analysis
```{r}
#Recaptures that I will analyze body mass for
recaptures_experiment_all_spp <- data_mass_wing_regressions %>%
  filter(experiment_bird == "yes") %>%
  mutate(bird_group = "all_spp")

#Create subset for GREH/VISA
recaptures_experiment_greh_visa <- recaptures_experiment_all_spp %>%
  filter(bird_species == "GREH" | bird_species == "VISA") %>%
  mutate(bird_group = "greh_visa")

#Combine subsets, add info about quantitative treatment
recaptures_experiment_for_analysis <- recaptures_experiment_all_spp %>%
  bind_rows(recaptures_experiment_greh_visa) %>%
  left_join(heto_removed) %>%
  mutate(log_calories_removed_per_ha = log(calories_removed_per_ha + 1)) %>%
  mutate(analysis = "body_mass",
         subanalysis = "body_mass") %>%
  select(analysis, subanalysis, bird_group, everything())
```

**Import functions**
```{r}
source("../../code/helper_functions/Modeling_helper_functions.R")
source("../../code/helper_functions/Plotting_helper_functions.R")
source("../../code/helper_functions/Extract_data_from_safely_and_quietly_lists.R")
source("../../code/helper_functions/Calculate_basic_summary_stats.R")
```

## Step 2: Using all capture data, plot log(mass) ~ log(wing length) for each species

Write functions
```{r}
#Subset data to each species
subset_data <- function(dataset, bird_species){
  
  data <- dataset %>%
    filter(bird_species == {{ bird_species }})
  
  return(data)
  
}

#test_data = subset_data(dataset = data_mass_wing_regressions, bird_species = "VISA")

#Create model for mass ~ wing length
create_model <- function(dataset){
  
  model <- lm(log_mass ~ log_wing_length, data = dataset)
  return(model)
  
}

#Add residuals and predicted values
augment_data <- function(dataset, model_object){
  
  data <- dataset %>%
    select(-log_mass, -log_wing_length) %>%
    cbind(broom::augment(model_object)) 

  return(data)
  
}

#Find outliers using cook's distance
identify_outliers <- function(dataset, sample_size){
  
  data <- dataset %>%
    rename(cooksd = ".cooksd") %>%
    mutate(cooksd_cutoff = 8/sample_size) %>% #Usually the cutoff is 4/n, but that was kicking out too much data
    mutate(cooksd_outlier = ifelse(cooksd > cooksd_cutoff, 1, 0))

  return(data)
  
}

#Plot mass ~ wing length
#Goal: identify outliers
make_plot_outlier_id <- function(dataset, bird_species, group_id){
  
  plot <- ggplot(aes(y = log_mass, x = log_wing_length), data = dataset) +
    geom_point(aes(colour = factor(get(group_id))), size = 2) +
    geom_line(aes(y = .fitted, x = log_wing_length), data = dataset) + #This is exactly the same as using geom_smooth() with method LM
    labs(title = bird_species, x = "Log(wing length)", y = "Log(mass)") +
    scale_color_manual(values = c("gray80", "#BA0022")) +
    theme_bw(base_size = 18) +
    theme(legend.position = "none")

  return(plot)
  
}

#Plot mass ~ wing length
#Goal: plot allometric equations
make_plot <- function(dataset, bird_species, intercept, slope, group_id){
  
  intercept = round(intercept, digits = 2)
  slope = round(slope, digits = 2)
  label = bquote("y ="~.(intercept)~"+"~.(slope)~"x") #Without parse(), it is a call.
  label_df <- data.frame(label = deparse(label))
  
  plot <- ggplot(aes(y = log_mass, x = log_wing_length), data = dataset) +
    geom_point(aes(colour = factor(get(group_id))), size = 2) +
    geom_line(aes(y = .fitted, x = log_wing_length), data = dataset) + #This is exactly the same as using geom_smooth() with method LM
    geom_text(aes(x = Inf, y = Inf, label = label), data = label_df, parse = TRUE, hjust = 1.25, vjust = 1.5, size = 6, colour = "black") + #Upper right
    labs(title = bird_species, x = "Ln(wing length)", y = "Ln(mass)") +
    scale_color_manual(values = c("gray80", "#BA0022")) +
    theme_bw(base_size = 18) +
    theme(legend.position = "none")

  return(plot)
  
}
```

Apply functions
```{r}
#1. Subset larger capture dataset to species of interest
#2. Only run regressions for species that have >30 observations
#3. Run regression
#4. Add residuals and predicted values from model, using broom::augment
mass_wing_regressions <- data_mass_wing_regressions %>%
  distinct(bird_species, bird_common_name) %>%
  mutate(data = map(bird_species, ~subset_data(dataset = data_mass_wing_regressions, bird_species = .))) %>% 
  mutate(num_rows = map(data, function(.){unlist(dim(.)[[1]])})) %>%
  filter(num_rows > 30) %>% #Remove species that don't have > 30 observations for regression
  mutate(mass_wing_model = map(data, create_model)) %>%
  mutate(data_augment = pmap(list(data, mass_wing_model), ~augment_data(dataset = ..1, model_object = ..2)))
```

Identify and remove outliers from regression
```{r}
#Identify outliers in mass ~ wing length regressions, with the goal of removing sketchy points (in lieu of going back to paper datasheets)
outliers <- mass_wing_regressions %>%
  mutate(data_augment_outliers = pmap(list(data_augment, num_rows), ~identify_outliers(dataset = ..1, sample_size = ..2))) %>%
  mutate(plot_outlier = pmap(list(data_augment_outliers, bird_common_name), ~make_plot_outlier_id(dataset = ..1, bird_species = ..2, group_id = "cooksd_outlier")))

#Sample plot of what outliers look like
outliers$plot_outlier[[1]]

#As data frame...
outliers_df <- bind_rows(outliers$data_augment_outliers) %>%
  filter(cooksd_outlier == 1)

#Are there any birds ID'ed as outliers that are in my recap dataset? NO.
check01 <- outliers_df %>%
  filter(bird_id %in% recaptures_experiment_for_analysis$bird_id)

#Remove outliers
data_mass_wing_regressions_without_outliers <- data_mass_wing_regressions %>%
  filter(!(bird_id %in% outliers_df$bird_id))
```

Re-run regressions using data without outliers
```{r, fig.height = 11.5, fig.width = 15}
mass_wing_regressions_without_outliers <- data_mass_wing_regressions %>%
  distinct(bird_species, bird_common_name) %>%
  mutate(data = map(bird_species, ~subset_data(dataset = data_mass_wing_regressions_without_outliers, bird_species = .))) %>% 
  mutate(num_rows = map(data, function(.){unlist(dim(.)[[1]])})) %>%
  filter(num_rows > 30) %>% #Remove species that don't have > 30 observations for regression
  mutate(mass_wing_model = map(data, create_model)) %>%
  mutate(mass_wing_intercept = map(mass_wing_model, function(.){coefficients(.)[[1]]})) %>%
  mutate(mass_wing_slope = map(mass_wing_model, function(.){coefficients(.)[[2]]})) %>%
  mutate(data_augment = pmap(list(data, mass_wing_model), ~augment_data(dataset = ..1, model_object = ..2))) %>%
  mutate(plot_recaps = pmap(list(data_augment, bird_common_name, mass_wing_intercept, mass_wing_slope), ~make_plot(dataset = ..1, bird_species = ..2, intercept = ..3, slope = ..4, group_id = "experiment_bird"))) %>%
  arrange(desc(as.numeric(num_rows))) #Re-order so that plots are ordered by amount of data

mass_wing_regressions_without_outliers$plot_recaps[[1]]
mass_wing_regressions_without_outliers$mass_wing_model[[1]] %>% summary()
mass_wing_regressions_without_outliers$num_rows[[1]]

#Combine plots of mass-wing regressions, for Supp Material
mass_wing_regressions_multiplot <- ggpubr::ggarrange(plotlist = mass_wing_regressions_without_outliers$plot_recaps)
mass_wing_regressions_multiplot

ggview::ggview(mass_wing_regressions_multiplot, height = 11.5, width = 15)
```

## Step 4: Calculate relative body mass
```{r}
#Get relative mass for birds in experiment (2016-2018)
relative_mass_for_analysis <- bind_rows(mass_wing_regressions_without_outliers$data_augment) %>%
  filter(experiment_bird == "yes") %>%
  rename(predicted_log_mass = ".fitted") %>%
  mutate(relative_mass = log_mass/predicted_log_mass) %>%
  select(-starts_with(".")) %>%
  select(-experiment_bird)

#Add relative mass to rest of data
data_for_analysis <- recaptures_experiment_for_analysis %>%
  left_join(relative_mass_for_analysis)
```

## Step 5: Analyze how relative body mass changes (or not) in response to *Heliconia* removal

**Prepare for analysis**
```{r}
#Create tables that will be filled in with models and model results
analysis = c("body_mass")
subanalysis = c("body_mass")
bird_group = c("all_spp", "greh_visa")
model_type = c("categorical", "quantitative")
model_name = c("control/treatment * pre/post", "calories removed/ha * pre/post")
treatment_variable = c("control_treatment", "log_calories_removed_per_ha")
model_types_names = data.frame(model_type = model_type, model_name = model_name, treatment_variable = treatment_variable)

base_table <- crossing(analysis, subanalysis, bird_group, model_type) %>% 
  left_join(model_types_names) %>%
  mutate(yvar = "relative_mass") %>%
  arrange(analysis, subanalysis, model_type, bird_group) %>%
  mutate(model_number = 1:length(yvar)) %>%
  mutate(title = paste(analysis, subanalysis, bird_group, sep = "_")) %>%
  mutate(subtitle = yvar) %>%
  select(model_number, analysis, yvar, subanalysis, everything())
```

**Filter master dataset to relevant sub-dataset & scale continuous predictors**
```{r}
subset_data <- function(dataset, analysis, subanalysis, bird_group){
  
  dataset <- dataset %>%
    filter(analysis == {{ analysis }}) %>%
    filter(subanalysis == {{ subanalysis}}) %>%
    filter(bird_group == {{ bird_group }}) %>%
    mutate(across(c(log_calories_removed_per_ha), ~c(scale(.)))) %>%
    mutate(exp_phase = factor(exp_phase, levels = c("pre", "post")))
      
  return(dataset)
  
}

#Create datasets (will add models in later step)
data <- base_table %>%
    mutate(data = pmap(list(analysis, subanalysis, bird_group), ~subset_data(dataset = data_for_analysis, analysis = ..1, subanalysis = ..2, bird_group = ..3)))
```

**Initial model fitting**
```{r, include = FALSE}
fit_data01 <- data_for_analysis %>%
  filter(analysis == "body_mass" & subanalysis == "body_mass" & bird_group == "all_spp")

#Can I include full random effect structure?
m1 <- glmmTMB(relative_mass ~ control_treatment * exp_phase + (1|patch/year_patch/bird_id), data = fit_data01, family = "gaussian", control = glmmTMBControl(optimizer=optim, optArgs=list(method="BFGS")))
m2 <- glmmTMB(relative_mass ~ control_treatment * exp_phase + (1|year_patch/bird_id), data = fit_data01, family = "gaussian", control = glmmTMBControl(optimizer=optim, optArgs=list(method="BFGS")))
m3 <- glmmTMB(relative_mass ~ control_treatment * exp_phase + (1|bird_id), data = fit_data01, family = "gaussian", control = glmmTMBControl(optimizer=optim, optArgs=list(method="BFGS")))

#No; run into convergence problems if try to include RE beyond bird_id

#Do assumptions for this model look OK?
performance::check_model(m3)

#Yes; look OK, but will check for all models later on.
```

**Fit models**
```{r}
create_model <- function(dataset, analysis, subanalysis, treatment_variable){

  model <- glmmTMB(as.formula(paste0("relative_mass ~ ", treatment_variable, " * exp_phase + (1|bird_id)")), data = dataset, na.action = na.omit, family = "gaussian")
      
  return(model)
  
}

#Create models
models <- data %>%
   mutate(model_quietly = pmap(list(data, analysis, subanalysis, treatment_variable), quietly(function(a, b, c, d){model <- create_model(dataset = a, analysis = b, subanalysis = c, treatment_variable = d)})),
          model = map(model_quietly, get_result),
          warning = map(model_quietly, get_warning), 
          convergence = map(model, check_convergence)) %>%
  mutate(xvar_table = map(model, ~data.frame(xvar = all.vars(terms(.))[-1])))

#Make sure there weren't any problems + check to make sure table of predictor variables is correct
models$warning
models$convergence
models$xvar_table
```

**Check assumptions**
```{r, fig.show = FALSE}
assumption_checks <- models %>%
    mutate(model_check = map(model, performance::check_model))

assumption_checks$model_check
```

**Get results together for export**
```{r}
results <- models %>%
  mutate(summary = map(model, summary),
         summary_tidy = map(model, broom.mixed::tidy),
         confint_wald = map(model, ~calculate_ci(., method_name = "wald"))) %>%
  mutate(num_obs = map(model, get_number_obs),
         num_levels = map(model, get_number_re_levels),
         num_reps = map(data, ~get_sample_size(data = ., vars = c("year", "patch", "control_treatment"), grouping_var = "control_treatment")))
```

## Step 6: Export results
```{r}
#Multiplot of log(mass) ~ log(wing length) regressions
path_name <- c("../../results/analysis/figures")
ggsave(filename = file.path(path_name, "FigS2_Allometric_regression_plots.png"), plot = mass_wing_regressions_multiplot, height = 11.5, width = 15, bg = "white")

#Responses to Heliconia removal
saveRDS(results, "../../results/analysis/rds_files/Body_mass_results.rds")
```
