==> code/analysis/09_visualizing_weights_and_percentage_resources_removed.Rmd <==
---
title: "Visualizing weights and resource availability"
output: html_document
knit: (function(inputFile, encoding) {
      out_dir <- "knitted_markdown_files";
      rmarkdown::render(inputFile,
                        encoding=encoding,
                        output_dir=file.path(dirname(inputFile), out_dir))})
---

**Goal:** Visualize relative sighting rates (i.e., weights used to tailor the resource estimate) and % calories removed. This is Figure 3 in the main text.

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE)
library(dplyr)
library(tidyr)
library(ggplot2)
library(purrr)
library(ggpubr)
```

## Step 1: Get data, get functions, specify figure dimensions

```{r}
weights <- read.csv("../../data/export/for_analysis/Weights.csv")
percentage_calories_removed_tailored <- read.csv("../../data/export/for_analysis/Percentage_calories_removed.csv")
source("../../code/helper_functions/Make_barplot.R")

#Recommended figure widths from journal (in mm)
column1x <- 80 #single-column
column2x <- 166 #double column

#Sizes for plot elements
size = 12
label_size = size
title_size = size * 1.15
shape_size = 1
line_width = 0.25
axis_size = 10

path_name <- c("../../results/analysis/figures")
```

## Step 2: Make boxplots of %  calories removed
```{r}
plot_theme = 
  theme_bw(base_size = size) +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.text.x = element_text(size = axis_size), #Need axis text to be small enough that labels do not overlap
        axis.line.x = element_line(colour = "black", size = line_width, linetype = "solid"),
        axis.line.y = element_line(colour = "black", size = line_width, linetype = "solid"),
        strip.background = element_blank(),
        strip.text.x = element_text(size = size),
        legend.position = "none")

data_for_boxplots <- percentage_calories_removed_tailored %>%
  filter(estimate_name == "low_estimate" | estimate_name == "high_estimate") %>%
  mutate(bird_group = factor(bird_group, levels = c("all_spp", "greh_visa"))) %>%
  mutate(estimate_name = factor(estimate_name, levels = c("low_estimate", "high_estimate"), labels = c("Low estimate", "High estimate")))

percentage_calories_removed_boxplots <-  data_for_boxplots %>%
  ggplot(data = ., aes(x = bird_group, y = percentage_removed)) +
    facet_grid(. ~ estimate_name, scales = "free_x") +
    geom_boxplot(position = position_nudge(x = 0, y = 0), width = 0.15, outlier.shape = NA, alpha = 1, fill = "grey80", linewidth = line_width) +
    geom_point(position = position_nudge(x = 0.15, y = 0), size = shape_size/2, colour = "black", alpha = 0.6, stroke = line_width) +
    labs(x = "Hummingbird group", y = "Calories removed (%)") +
    stat_summary(fun = mean, geom = "point", shape = 23, size = shape_size, color = "black", fill = "black") +
    scale_y_continuous(breaks = c(0, 20, 40, 60, 80, 100)) +
    plot_theme +
    scale_x_discrete(labels = c("All species", expression(italic(Heliconia)~"specialists")))

percentage_calories_removed_boxplots
```

## Step 3: Make barplots of relative sighting rates (weights)
```{r}
data_for_barplots <- weights %>%
  mutate(plant_scientific_name = ifelse(plant_scientific_name == "Drymonia sp.", "Drymonia", plant_scientific_name)) %>% 
  mutate(plant_scientific_name = ifelse(plant_scientific_name == "Renealmia sp.", "Renealmia", plant_scientific_name)) %>% 
  mutate(heto_color = ifelse(plant_species == "HETO", "yes", "no"))
```

```{r}
#All species 
weights_barplot_all_spp <- data_for_barplots %>%
  filter(bird_group == "all_spp") %>% 
  make_barplot(xvar = "plant_scientific_name", yvar = "weight", xlabel = "Plant species", ylabel = "Relative visitation rate", highlight_column = "heto_color", titlelabel = "All hummingbird species", flip_axes = FALSE, text_size = size)

weights_barplot_all_spp

#GREH/VISA
weights_barplot_gv <- data_for_barplots %>%
  filter(bird_group == "greh_visa") %>% 
  make_barplot(xvar = "plant_scientific_name", yvar = "weight", xlabel = "Plant species", ylabel = "Relative visitation rate", highlight_column = "heto_color", titlelabel = "Heliconia specialists", flip_axes = FALSE, text_size = size)

weights_barplot_gv
```

## Step 4: Combine weights and % calories removed boxplots into one multi-panel figure

```{r}
#Add spacing around boxplots
boxplots_for_multiplot <- ggarrange(NULL, percentage_calories_removed_boxplots, NULL, ncol = 3, nrow = 1, widths = c(0.10, 1, 0.10))

weights_and_percent_multiplot <- ggarrange(weights_barplot_all_spp, weights_barplot_gv, boxplots_for_multiplot, ncol = 1, nrow = 3, heights = c(1, 1, 1), labels = c("(a)", "(b)", "(c)"), font.label = list(size = label_size, face = "plain"))

ggview::ggview(weights_and_percent_multiplot, width = column2x, height = column2x*1.5, units = "mm")

ggsave(file.path(path_name, "Fig3_Weights_Percentage_calories.png"), weights_and_percent_multiplot, width = column2x, height = column2x*1.5, units = "mm", bg = "white", dpi = 600)
```
